feat: integrate Stripe payment processing and update order management

+++ cnm_bookhub_be/web/api/books/schema.py
@@ -1,4 +1,5 @@
 import uuid
+
 from pydantic import BaseModel, ConfigDict
 
 
@@ -6,7 +7,7 @@ class BookDTO(BaseModel):
     id: uuid.UUID
     title: str
     author: str
-    price: float
+    price: int
     stock: int
     image_urls: str | None
     description: str | None
@@ -19,7 +20,7 @@ class BookDTO(BaseModel):
 class BookCreateDTO(BaseModel):
     title: str
     author: str
-    price: float
+    price: int
     stock: int = 0
     image_urls: str | None = None
     description: str | None = None
@@ -30,18 +31,19 @@ class BookCreateDTO(BaseModel):
 class BookUpdateDTO(BaseModel):
     title: str | None = None
     author: str | None = None
-    price: float | None = None
+    price: int | None = None
     stock: int | None = None
     image_urls: str | None = None
     description: str | None = None
     more_info: dict | None = None
     category_id: int | None = None
 
+
 class BooksByCategoryDTO(BaseModel):
     id: uuid.UUID
     title: str
     author: str
-    price: float
+    price: int
     stock: int
     image_urls: str | None = None
     description: str | None = None
